Fix event queue IDs using wrong address size in 64-bit

+++ src/game/server/eventqueue.h
@@ -41,8 +41,8 @@ class CEventQueue
 public:
 	// pushes an event into the queue, targeting a string name (m_iName), or directly by a pointer
 #ifdef MAPBASE_VSCRIPT
-	int      AddEvent( const char  *target, const char *action, variant_t Value, float fireDelay, CBaseEntity *pActivator, CBaseEntity *pCaller, int outputID = 0 );
-	int      AddEvent( CBaseEntity *target, const char *action, variant_t Value, float fireDelay, CBaseEntity *pActivator, CBaseEntity *pCaller, int outputID = 0 );
+	intp      AddEvent( const char  *target, const char *action, variant_t Value, float fireDelay, CBaseEntity *pActivator, CBaseEntity *pCaller, int outputID = 0 );
+	intp      AddEvent( CBaseEntity *target, const char *action, variant_t Value, float fireDelay, CBaseEntity *pActivator, CBaseEntity *pCaller, int outputID = 0 );
 #else
 	void     AddEvent( const char  *target, const char *action, variant_t Value, float fireDelay, CBaseEntity *pActivator, CBaseEntity *pCaller, int outputID = 0 );
 	void     AddEvent( CBaseEntity *target, const char *action, variant_t Value, float fireDelay, CBaseEntity *pActivator, CBaseEntity *pCaller, int outputID = 0 );
@@ -73,8 +73,8 @@ public:
 
 #ifdef MAPBASE_VSCRIPT
 	void CancelEventsByInput( CBaseEntity *pTarget, const char *szInput );
-	bool RemoveEvent( int event );
-	float GetTimeLeft( int event );
+	bool RemoveEvent( intp event );
+	float GetTimeLeft( intp event );
 #endif // MAPBASE_VSCRIPT
 
 private:
